Conflicts:
	opencell-admin/web/.metadata/src/main/webapp/WEB-INF/faces-config.pageflow
	opencell-admin/web/src/main/webapp/layout/navigationItems.xhtml
	opencell-api-dto/src/main/java/org/meveo/api/dto/payment/AccountOperationDto.java
	opencell-api/src/main/java/org/meveo/api/OccTemplateApi.java
	opencell-api/src/main/java/org/meveo/api/payment/AccountOperationApi.java
	opencell-api/src/main/java/org/meveo/api/payment/RefundApi.java
	opencell-model/src/main/db_resources/changelog/current/data.xml
	opencell-model/src/main/db_resources/changelog/current/structure.xml
	opencell-model/src/main/java/org/meveo/model/payments/AccountOperation.java
	opencell-tests/api/soapui/Full-WS-CRUD-Test-Entities.xml

+++ opencell-admin/ejbs/src/main/java/org/meveo/service/base/PersistenceService.java
@@ -69,8 +69,8 @@ import org.meveo.model.filter.Filter;
 import org.meveo.service.base.local.IPersistenceService;
 import org.meveo.service.crm.impl.CustomFieldInstanceService;
 import org.meveo.service.index.ElasticClient;
-import org.meveo.util.MeveoJpa;
-import org.meveo.util.MeveoJpaForJobs;
+import org.meveo.util.MeveoJpaForMultiTenancy;
+import org.meveo.util.MeveoJpaForMultiTenancyForJobs;
 
 /**
  * Generic implementation that provides the default implementation for persistence methods declared in the {@link IPersistenceService} interface.
@@ -89,11 +89,11 @@ public abstract class PersistenceService<E extends IEntity> extends BaseService 
     private ParamBean paramBean = ParamBean.getInstance();
 
     @Inject
-    @MeveoJpa
+    @MeveoJpaForMultiTenancy
     private EntityManager em;
 
     @Inject
-    @MeveoJpaForJobs
+    @MeveoJpaForMultiTenancyForJobs
     private EntityManager emfForJobs;
 
     @Inject
@@ -101,9 +101,6 @@ public abstract class PersistenceService<E extends IEntity> extends BaseService 
 
     @Inject
     private Conversation conversation;
-    //
-    // @Resource
-    // protected TransactionSynchronizationRegistry txReg;
 
     @Inject
     @Created
@@ -295,28 +292,31 @@ public abstract class PersistenceService<E extends IEntity> extends BaseService 
     @Override
     public void remove(E entity) throws BusinessException {
         log.debug("start of remove {} entity (id={}) ..", getEntityClass().getSimpleName(), entity.getId());
-        getEntityManager().remove(entity);
-        if (entity instanceof BaseEntity && entity.getClass().isAnnotationPresent(ObservableEntity.class)) {
-            entityRemovedEventProducer.fire((BaseEntity) entity);
-        }
-        // getEntityManager().flush();
+        entity = findById((Long) entity.getId());
+        if (entity != null) {
+            getEntityManager().remove(entity);
+            if (entity instanceof BaseEntity && entity.getClass().isAnnotationPresent(ObservableEntity.class)) {
+                entityRemovedEventProducer.fire((BaseEntity) entity);
+            }
+            // getEntityManager().flush();
 
-        // Remove entity from Elastic Search
-        if (BusinessEntity.class.isAssignableFrom(entity.getClass())) {
-            elasticClient.remove((BusinessEntity) entity);
-        }
+            // Remove entity from Elastic Search
+            if (BusinessEntity.class.isAssignableFrom(entity.getClass())) {
+                elasticClient.remove((BusinessEntity) entity);
+            }
 
-        // Remove custom field values from cache if applicable
-        if (entity instanceof ICustomFieldEntity) {
-            customFieldInstanceService.removeCFValues((ICustomFieldEntity) entity);
-        }
+            // Remove custom field values from cache if applicable
+            if (entity instanceof ICustomFieldEntity) {
+                customFieldInstanceService.removeCFValues((ICustomFieldEntity) entity);
+            }
 
-        if (entity instanceof IImageUpload) {
-            try {
-                ImageUploadEventHandler<E> imageUploadEventHandler = new ImageUploadEventHandler<E>(appProvider);
-                imageUploadEventHandler.deleteImage(entity);
-            } catch (IOException e) {
-                log.error("Failed deleting image file");
+            if (entity instanceof IImageUpload) {
+                try {
+                    ImageUploadEventHandler<E> imageUploadEventHandler = new ImageUploadEventHandler<E>(appProvider);
+                    imageUploadEventHandler.deleteImage(entity);
+                } catch (IOException e) {
+                    log.error("Failed deleting image file");
+                }
             }
         }
 
@@ -644,7 +644,7 @@ public abstract class PersistenceService<E extends IEntity> extends BaseService 
 
         return refreshedEntities;
     }
-    
+
     /**
      * Creates query to filter entities according data provided in pagination configuration.
      * 
@@ -1026,19 +1026,6 @@ public abstract class PersistenceService<E extends IEntity> extends BaseService 
         return false;
     }
 
-    public EntityManager getEntityManager() {
-        EntityManager result = emfForJobs;
-        if (conversation != null) {
-            try {
-                conversation.isTransient();
-                result = em;
-            } catch (Exception e) {
-            }
-        }
-
-        return result;
-    }
-
     public void updateAudit(E e) {
         if (e instanceof IAuditable) {
             ((IAuditable) e).updateAudit(currentUser);
@@ -1063,5 +1050,18 @@ public abstract class PersistenceService<E extends IEntity> extends BaseService 
             }
         }
         return q.getResultList();
+    }
+
+    public EntityManager getEntityManager() {
+        if (conversation != null) {
+            try {
+                conversation.isTransient();
+                return em;
+            } catch (Exception e) {
+                return emfForJobs;
+            }
+        }
+
+        return emfForJobs;
     }
 }
